+++ internal/install/types/types.go
@@ -376,6 +376,8 @@ type OpenInstallationRecipe struct {
 	// NRQL the newrelic-cli uses to validate this recipe
 	// is successfully sending data to New Relic
 	ValidationNRQL NRQL `json:"validationNrql,omitempty"`
+	// validation url to validate with infra health endpoint
+	ValidationURL string `json:"validationUrl,omitempty"`
 }
 
 // OpenInstallationRecipeInputVariable - Recipe input variable prompts displayed to the user prior to execution